Changed the id in ingredient_type table to type 'serial' which is the PostgreSQL of auto-generating the id key. Also fixed a typo in the view and converted the ingredient layout to use a table

+++ recipe-scavenger/database/schemas/public/tables/ingredient_type.sql
@@ -4,7 +4,7 @@
 
 CREATE TABLE ingredient_type
 (
-  id bigint NOT NULL,
+  id serial NOT NULL,
   "name" text NOT NULL,
   base_uom character varying(25) NOT NULL,
   CONSTRAINT ingredient_type_pkey PRIMARY KEY (id)
